Split bot functionality tests corresponding to the functionality modules
